add plots
