+++ simwise/__init__.py
@@ -3,5 +3,5 @@ from simwise.satellite import SatelliteState, SatelliteParams
 from simwise.dynamics import state_dot, attitude_dot, orbit_dot
 from simwise.torques import gravity_gradient
 from simwise.forces import j2
-from simwise.utils import mjd, gmst, date2mjd
+from simwise.utils import mjd, gmst, date2mjd, build_attitude_fig, build_orbit_fig
 from simwise.propagator import propagate
