+++ simwise/utils/__init__.py
@@ -2,3 +2,4 @@ from simwise.utils.transforms import ecef2blh, eci2ecef, ecef2eci
 from simwise.utils.time import gmst, date2mjd
 from simwise.utils.time import mjd, gmst, date2mjd
 from simwise.utils.orbital_elements import state2coe, coe2state
+from simwise.utils.plots import build_attitude_fig, build_orbit_fig
